refactor github into another file
